+++ src/main/java/me/pajic/simple_music_control/gui/NowPlayingWidget.java
@@ -2,7 +2,6 @@ package me.pajic.simple_music_control.gui;
 
 import me.pajic.simple_music_control.SMC;
 import me.pajic.simple_music_control.util.ModUtil;
-import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.client.gui.screens.PauseScreen;
@@ -17,79 +16,78 @@ public class NowPlayingWidget {
     private static final Minecraft MC = Minecraft.getInstance();
     private static SoundInstance soundInstance = null;
     private static Component trackName;
-    private static final Component note = Component.literal("♫");
-    private static float timer = 0;
-    private static float toggleTimer = 0;
+    private static final Component NOTE = Component.literal("♫");
+    private static long timer = 0;
+    private static long toggleTimer = 0;
+	private static long previousTickMillis = System.currentTimeMillis();
     private static boolean centered = true;
 
-    public static void render(GuiGraphicsExtractor guiGraphics, DeltaTracker deltaTracker) {
-        if (SMC.CONFIG.nowPlayingWidget.get() && !MC.options.hideGui && isMusicOn() && soundInstance != null && timer > 0) {
-            if (soundInstance.getSound() == null) {
-                renderConditionalText(guiGraphics, Component.translatable("gui.simple_music_control.soundSystemTrippedFellAndExploded"), Color.WHITE.getRGB());
-            } else if (!(MC.level != null && MC.screen != null)) {
-                renderConditionalTextWithFade(trackName, guiGraphics, deltaTracker, true);
-                if (!MC.getSoundManager().isActive(soundInstance)) soundInstance = null;
-            }
-        } else if (toggleTimer > 0) {
-            renderConditionalTextWithFade(ModUtil.globalPause ?
-                    Component.translatable("gui.simple_music_control.pause") :
-                    Component.translatable("gui.simple_music_control.resume"),
-                    guiGraphics, deltaTracker, false);
-            if (ModUtil.globalPause) soundInstance = null;
-        }
+    public static void render(GuiGraphicsExtractor guiGraphics) {
+		if (!(MC.screen instanceof PauseScreen)) {
+		    if (SMC.CONFIG.nowPlayingWidget.get() && !MC.options.hideGui && isMusicOn() && soundInstance != null && timer > 0) {
+				if (soundInstance.getSound() == null) {
+					renderConditionalText(guiGraphics, Component.translatable("gui.simple_music_control.soundSystemTrippedFellAndExploded"), Color.WHITE.getRGB(), false, 0);
+				} else if (!(MC.level != null && MC.screen != null)) {
+					renderConditionalTextWithFade(trackName, guiGraphics, true);
+					if (!MC.getSoundManager().isActive(soundInstance)) soundInstance = null;
+				}
+			} else if (toggleTimer > 0) {
+				renderConditionalTextWithFade(ModUtil.globalPause ?
+								Component.translatable("gui.simple_music_control.pause") :
+								Component.translatable("gui.simple_music_control.resume"),
+						guiGraphics, false);
+				if (ModUtil.globalPause) soundInstance = null;
+			}
+	    }
+		previousTickMillis = System.currentTimeMillis();
     }
 
     public static void displayWidget(SoundInstance sound) {
-        if (!(MC.screen instanceof PauseScreen)) {
-            soundInstance = sound;
-            if (soundInstance != null && soundInstance.getSound() != null) {
-                trackName = Component.translatable(soundInstance.getSound().getLocation().toShortLanguageKey().replace("/", "."));
-                centered = MC.level != null;
-                timer = SMC.CONFIG.nowPlayingWidgetDuration.get() * 20;
-            }
-        }
+		soundInstance = sound;
+		if (soundInstance != null && soundInstance.getSound() != null) {
+			trackName = Component.translatable(soundInstance.getSound().getLocation().toShortLanguageKey().replace("/", "."));
+			centered = MC.level != null;
+			timer = SMC.CONFIG.nowPlayingWidgetDuration.get() * 1000L;
+		}
     }
 
     public static void displayToggleNotification() {
         if (!(MC.screen instanceof PauseScreen)) {
             centered = MC.level != null;
-            toggleTimer = SMC.CONFIG.nowPlayingWidgetDuration.get() * 20;
+            toggleTimer = SMC.CONFIG.nowPlayingWidgetDuration.get() * 1000L;
         }
     }
 
     public static void displayPauseScreenWidget(GuiGraphicsExtractor guiGraphics) {
         if (SMC.CONFIG.showNowPlayingWidgetInPauseMenu.get() && isMusicOn() && soundInstance != null && soundInstance.getSound() != null) {
-            renderCornerText(guiGraphics, true);
+            renderCornerText(guiGraphics, true, Color.WHITE.getRGB(), MusicNoteColorManager.musicNoteColor);
         }
 		if (!MC.getSoundManager().isActive(soundInstance)) soundInstance = null;
     }
 
-    private static void renderConditionalText(GuiGraphicsExtractor guiGraphics, Component text, int textColor, int noteColor) {
+    private static void renderConditionalText(GuiGraphicsExtractor guiGraphics, Component text, int textColor, boolean note, int noteColor) {
         if (centered) {
             renderActionBarText(text, guiGraphics, 0, textColor);
-            renderActionBarText(note, guiGraphics, MC.font.width(trackName) / 2 + 7, noteColor);
-            renderActionBarText(note, guiGraphics, -MC.font.width(trackName) / 2 - 7, noteColor);
+			if (note) {
+		        renderActionBarText(NOTE, guiGraphics, MC.font.width(trackName) / 2 + 7, noteColor);
+				renderActionBarText(NOTE, guiGraphics, -MC.font.width(trackName) / 2 - 7, noteColor);
+	        }
         }
-        else renderCornerText(guiGraphics, false);
+        else renderCornerText(guiGraphics, false, textColor, noteColor);
     }
 
-    private static void renderConditionalText(GuiGraphicsExtractor guiGraphics, Component text, int textColor) {
-        if (centered) renderActionBarText(text, guiGraphics, 0, textColor);
-		else renderCornerText(guiGraphics, false);
-    }
-
-    private static void renderConditionalTextWithFade(Component text, GuiGraphicsExtractor guiGraphics, DeltaTracker deltaTracker, boolean note) {
-        float f = note ? timer - deltaTracker.getRealtimeDeltaTicks() : toggleTimer - deltaTracker.getRealtimeDeltaTicks();
-        int i = (int) (f * 255.0F / 20.0F);
+    private static void renderConditionalTextWithFade(Component text, GuiGraphicsExtractor guiGraphics, boolean note) {
+		long delta = System.currentTimeMillis() - previousTickMillis;
+        long t = note ? timer - delta : toggleTimer - delta;
+        int i = (int) (t * 255L / 1000L);
         if (i > 255) i = 255;
         if (i > 8) {
             int textColor = ARGB.color(i, Color.WHITE.getRGB());
             int noteColor = ARGB.color(i, MusicNoteColorManager.musicNoteColor);
-            if (note) renderConditionalText(guiGraphics, text, textColor, noteColor);
-            else renderConditionalText(guiGraphics, text, textColor);
+            renderConditionalText(guiGraphics, text, textColor, note, noteColor);
         }
-        if (note) timer -= deltaTracker.getRealtimeDeltaTicks();
-        else toggleTimer -= deltaTracker.getRealtimeDeltaTicks();
+        if (note) timer -= delta;
+        else toggleTimer -= delta;
     }
 
     private static void renderActionBarText(Component text, GuiGraphicsExtractor guiGraphics, int xOffset, int color) {
@@ -101,39 +99,31 @@ public class NowPlayingWidget {
         );
     }
 
-    private static void renderCornerText(GuiGraphicsExtractor guiGraphics, boolean pauseMenu) {
-        if (pauseMenu) {
-            switch (SMC.CONFIG.pauseWidgetPosition.get()) {
-                case TOP_LEFT -> {
-                    guiGraphics.text(MC.font, note, 4, 4, MusicNoteColorManager.musicNoteColor);
-                    guiGraphics.text(MC.font, trackName, 15, 4, Color.WHITE.getRGB());
-                }
-                case TOP_RIGHT -> {
-                    guiGraphics.text(MC.font, note, MC.getWindow().getGuiScaledWidth() - 12, 4, MusicNoteColorManager.musicNoteColor);
-                    guiGraphics.text(MC.font, trackName, MC.getWindow().getGuiScaledWidth() - MC.font.width(trackName) - 16, 4, Color.WHITE.getRGB());
-                }
-                case BOTTOM_LEFT -> {
-                    guiGraphics.text(MC.font, note, 4, MC.getWindow().getGuiScaledHeight() - 11, MusicNoteColorManager.musicNoteColor);
-                    guiGraphics.text(MC.font, trackName, 15, MC.getWindow().getGuiScaledHeight() - 11, Color.WHITE.getRGB());
-                }
-                case BOTTOM_RIGHT -> {
-                    guiGraphics.text(MC.font, note, MC.getWindow().getGuiScaledWidth() - 12, MC.getWindow().getGuiScaledHeight() - 11, MusicNoteColorManager.musicNoteColor);
-                    guiGraphics.text(MC.font, trackName, MC.getWindow().getGuiScaledWidth() - MC.font.width(trackName) - 16, MC.getWindow().getGuiScaledHeight() - 11, Color.WHITE.getRGB());
-                }
-            }
-        } else {
-            switch (SMC.CONFIG.pauseWidgetPosition.get()) {
-                case TOP_LEFT, BOTTOM_LEFT -> {
-                    guiGraphics.text(MC.font, note, 4, 4, MusicNoteColorManager.musicNoteColor);
-                    guiGraphics.text(MC.font, trackName, 15, 4, Color.WHITE.getRGB());
-                }
-                case TOP_RIGHT, BOTTOM_RIGHT -> {
-                    guiGraphics.text(MC.font, note, MC.getWindow().getGuiScaledWidth() - 12, 4, MusicNoteColorManager.musicNoteColor);
-                    guiGraphics.text(MC.font, trackName, MC.getWindow().getGuiScaledWidth() - MC.font.width(trackName) - 16, 4, Color.WHITE.getRGB());
-                }
-            }
-        }
+    private static void renderCornerText(GuiGraphicsExtractor guiGraphics, boolean pauseMenu, int textColor, int noteColor) {
+        if (pauseMenu) renderCornerText(SMC.CONFIG.pauseWidgetPosition.get(), guiGraphics, textColor, noteColor);
+		else renderCornerText(SMC.CONFIG.titleWidgetPosition.get(), guiGraphics, textColor, noteColor);
     }
+
+	private static void renderCornerText(WidgetPosition pos, GuiGraphicsExtractor guiGraphics, int textColor, int noteColor) {
+		switch (pos) {
+			case TOP_LEFT -> {
+				guiGraphics.text(MC.font, NOTE, 4, 4, noteColor);
+				guiGraphics.text(MC.font, trackName, 15, 4, textColor);
+			}
+			case TOP_RIGHT -> {
+				guiGraphics.text(MC.font, NOTE, MC.getWindow().getGuiScaledWidth() - 12, 4, noteColor);
+				guiGraphics.text(MC.font, trackName, MC.getWindow().getGuiScaledWidth() - MC.font.width(trackName) - 16, 4, textColor);
+			}
+			case BOTTOM_LEFT -> {
+				guiGraphics.text(MC.font, NOTE, 4, MC.getWindow().getGuiScaledHeight() - 11, noteColor);
+				guiGraphics.text(MC.font, trackName, 15, MC.getWindow().getGuiScaledHeight() - 11, textColor);
+			}
+			case BOTTOM_RIGHT -> {
+				guiGraphics.text(MC.font, NOTE, MC.getWindow().getGuiScaledWidth() - 12, MC.getWindow().getGuiScaledHeight() - 11, noteColor);
+				guiGraphics.text(MC.font, trackName, MC.getWindow().getGuiScaledWidth() - MC.font.width(trackName) - 16, MC.getWindow().getGuiScaledHeight() - 11, textColor);
+			}
+		}
+	}
 
     private static boolean isMusicOn() {
         return MC.options.getSoundSourceVolume(SoundSource.MASTER) > 0 && MC.options.getSoundSourceVolume(SoundSource.MUSIC) > 0;