+++ src/main/java/me/pajic/simple_music_control/gui/WidgetPosition.java
@@ -3,11 +3,11 @@ package me.pajic.simple_music_control.gui;
 import me.fzzyhmstrs.fzzy_config.util.EnumTranslatable;
 import org.jetbrains.annotations.NotNull;
 
-public enum PauseMenuWidgetPosition implements EnumTranslatable {
+public enum WidgetPosition implements EnumTranslatable {
     TOP_LEFT, TOP_RIGHT, BOTTOM_LEFT, BOTTOM_RIGHT;
 
 	@Override
 	@NotNull public String prefix() {
-		return "simple_music_control.config.pauseWidgetPosition";
+		return "simple_music_control.config.widgetPosition";
 	}
 }